fix: align waste reminder sync typing and registration activation
- aligned `locationData` typing with `WasteReminderServerSyncPayload` to avoid mismatched optional fields in the server sync flow
- extracted reminder registration builders into a shared helper to centralize the logic and reduce duplication in `WasteCollectionSettingsScreen`
- based activation of server sync registrations on `notificationSettings[typeKey]` so enabled reminders stayed active even when nested type settings had not been hydrated consistently
- added regression tests for registration activation and server sync setup to catch the reviewed edge cases and protect the refactor
