+++ src/screens/WasteCollectionSettingsScreen.tsx
@@ -41,6 +41,11 @@ import {
   WrapperVertical
 } from '../components';
 import { colors, consts, device, Icon, normalize, texts } from '../config';
+import {
+  buildDefaultReminderSettingsByType,
+  buildReminderServerSyncRegistrations,
+  mergeReminderSettingsWithDefaults
+} from '../helpers/wasteReminderRegistrationHelper';
 import { formatTime, saveWasteReminderSettings, storageHelper } from '../helpers';
 import { formatWasteReminderTime } from '../helpers/wasteReminderTimeHelper';
 import {
@@ -1006,32 +1011,6 @@ export const WasteCollectionSettingsScreen = () => {
 };
 /* eslint-enable complexity */
 
-const buildDefaultReminderSettingsByType = (
-  usedTypes: WasteTypeData
-): WasteReminderSettingsByType =>
-  Object.fromEntries(
-    Object.entries(usedTypes).map(([typeKey, wasteType]) => {
-      const normalizedSlots = normalizePushReminderSlots(wasteType);
-
-      return [
-        typeKey,
-        {
-          enabled: normalizedSlots.slots.length > 0,
-          reminders: Object.fromEntries(
-            normalizedSlots.slots.map((slot) => [
-              slot.id,
-              {
-                enabled: true,
-                leadDays: slot.defaultLeadDays,
-                time: '09:00'
-              }
-            ])
-          )
-        }
-      ];
-    })
-  );
-
 const buildReminderSettingsFromRegistrations = (
   usedTypes: WasteTypeData,
   registrations: Array<WasteReminderRegistration & { active?: boolean }>
@@ -1079,63 +1058,6 @@ const buildReminderSettingsFromServerSettings = (
       time: formatDateAsReminderTime(new Date(setting.notify_at)),
       typeKey: setting.notify_for_waste_type
     }))
-  );
-
-const buildReminderServerSyncRegistrations = (
-  reminderSettingsByType: WasteReminderSettingsByType,
-  usedTypes: WasteTypeData,
-  selectedTypeKeys: string[],
-  notificationSettings: { [key: string]: boolean }
-): WasteReminderServerSyncRegistration[] => {
-  const selectedTypes = new Set(selectedTypeKeys);
-  const completeReminderSettingsByType = mergeReminderSettingsWithDefaults(
-    buildDefaultReminderSettingsByType(usedTypes),
-    reminderSettingsByType
-  );
-
-  return Object.entries(completeReminderSettingsByType).flatMap(([typeKey, typeSetting]) => {
-    if (!selectedTypes.has(typeKey)) {
-      return [];
-    }
-
-    const isTypeActive = !!notificationSettings[typeKey] && typeSetting.enabled;
-
-    return Object.entries(typeSetting.reminders).map(([slotId, slotSetting]) => ({
-      active: isTypeActive && slotSetting.enabled,
-      leadDays: slotSetting.leadDays,
-      slotId,
-      storeId: slotSetting.storeId,
-      time: slotSetting.time,
-      typeKey
-    }));
-  });
-};
-
-const mergeReminderSettingsWithDefaults = (
-  defaultSettings: WasteReminderSettingsByType,
-  reminderSettingsByType: WasteReminderSettingsByType
-): WasteReminderSettingsByType =>
-  Object.fromEntries(
-    Object.entries(defaultSettings).map(([typeKey, defaultTypeSetting]) => {
-      const typeSetting = reminderSettingsByType[typeKey];
-
-      return [
-        typeKey,
-        {
-          ...defaultTypeSetting,
-          ...typeSetting,
-          reminders: Object.fromEntries(
-            Object.entries(defaultTypeSetting.reminders).map(([slotId, defaultSlotSetting]) => [
-              slotId,
-              {
-                ...defaultSlotSetting,
-                ...typeSetting?.reminders[slotId]
-              }
-            ])
-          )
-        }
-      ];
-    })
   );
 
 const buildStoredSettingsFromLocalPayload = (
